Create Day_7-guess the word
Randomly choose a word from the word_list and assign it to a variable called chosen_word. Then print it.
Ask the user to guess a letter and assign their answer to a variable called guess. Make the String stored in guess lowercase.
Check if the letter the user guessed guess is one of the letters in the chosen_word. Loop through each of the letters in the chosen_word and print "Right" if the letter is a match, "Wrong" if it's not.